adding some features and new modules

+++ ThackTech/Pipelines/PipelineModules/SortBam.py
@@ -1,13 +1,18 @@
 import os
 import shutil 
 import subprocess
-from ThackTech.Pipelines import PipelineModule
+from ThackTech.Pipelines import PipelineModule, ModuleParameter
+from ThackTech.Pipelines.FileInfo import FileInfo, FileContext
 
 
 class SortBam(PipelineModule):
 	
 	def __init__(self, **kwargs):
 		super(SortBam, self).__init__('SortBam', 'Sort BAM file', **kwargs)
+		
+		self.add_parameter(ModuleParameter('sort_name', bool, False, desc="Sort by read name rather than coordinate."))
+		self.add_parameter(ModuleParameter('overwrite', bool, True, desc="Overwrite the source, non-sorted, BAM."))
+		
 		self._name_resolver('alignments')
 	#end __init__()
 
@@ -23,9 +28,11 @@ class SortBam(PipelineModule):
 		
 		sort_cmd = ['samtools', 'sort', '-@', str(self.processors), bam, sorted_bam]
 		
+		if self.get_parameter_value('sort_name'):
+			sort_cmd.insert(2, '-n')
+		
 		#Sort our BAM file by genomic location: -@ => multithreading!
 		cxt.log.write('-> Sorting BAM "%s" (using %d processor%s)...\n' % (bam, self.processors, ('s' if self.processors > 1 else '')))
-		cxt.log.write("-> "+subprocess.check_output('samtools 2>&1 | grep Version', shell=True, stderr=subprocess.STDOUT)+"")
 		cxt.log.write("..............................................\n")
 		cxt.log.write(" ".join(sort_cmd))
 		cxt.log.write("\n..............................................\n")
@@ -33,8 +40,11 @@ class SortBam(PipelineModule):
 		
 		self._run_subprocess(sort_cmd, stderr=subprocess.STDOUT, stdout=cxt.log)
 		
-		#move the sorted BAM to overwrite the unsorted BAM
-		os.remove(bam) #first remove the origional unsorted bam, otherwise the move will choke!
-		shutil.move(sorted_bam+'.bam', bam) #mv "${dest}${alnname}_sorted.bam" "${dest}${alnname}.bam"
+		if self.get_parameter_value('overwrite'):
+			#move the sorted BAM to overwrite the unsorted BAM
+			os.remove(bam) #first remove the origional unsorted bam, otherwise the move will choke!
+			shutil.move(sorted_bam+'.bam', bam) #mv "${dest}${alnname}_sorted.bam" "${dest}${alnname}.bam"
+		else:
+			return [FileInfo(sorted_bam+'.bam', FileContext.from_module_context(cxt, "sorted_bam"))]
 	#end run()
 #end class GeneratePseudoreplicates
